fix: retrieve amountIn
