+++ src/features/gift/components/GiftTakerSuccessScreen.tsx
@@ -1,8 +1,11 @@
 import { Check as CheckIcon } from "@phosphor-icons/react"
 import { useQuery } from "@tanstack/react-query"
+import {
+  computeTotalBalanceDifferentDecimals,
+  getUnderlyingBaseTokenInfos,
+} from "src/utils/tokenUtils"
 import { CopyButton } from "../../../components/IntentCard/CopyButton"
 import { waitForIntentSettlement } from "../../../services/intentService"
-import type { TokenValue } from "../../../types/base"
 import { assert } from "../../../utils/assert"
 import type { GiftTerms } from "../utils/deriveGiftTerms"
 import { GiftStrip } from "./GiftStrip"
@@ -12,12 +15,18 @@ const NEAR_EXPLORER = "https://nearblocks.io"
 export function GiftTakerSuccessScreen({
   giftTerms,
   intentHashes,
-  amountIn,
 }: {
   giftTerms: GiftTerms
   intentHashes: string[]
-  amountIn: TokenValue
 }) {
+  const amountIn = computeTotalBalanceDifferentDecimals(
+    getUnderlyingBaseTokenInfos(giftTerms.tokenIn),
+    giftTerms.tokenDiff,
+    { strict: false }
+  )
+
+  assert(amountIn != null)
+
   const intentStatus = useQuery({
     queryKey: ["intents_status", intentHashes],
     queryFn: async ({ signal }) => {
